From Scratch Code
This is my first attempt at the website from scratch. The look of it is super rough, but I got all the navigation working and there are only a few more things I need to add to each page in terms of HTML and then I'll be able to start styling.